Rename files, change macro logic
