+++ src/corecel/sys/TraceCounter.hh
@@ -13,22 +13,29 @@
 #include "corecel/Config.hh"
 
 #include "ScopedProfiling.hh"
-#if CELERITAS_USE_PERFETTO || CELERITAS_USE_CUDA
-#    include "detail/TraceCounterImpl.hh"
-#endif
+
+#include "detail/TraceCounterImpl.hh"
 
 namespace celeritas
 {
-#if CELERITAS_USE_PERFETTO || CELERITAS_USE_CUDA
 //---------------------------------------------------------------------------//
-// Simple tracing counter
+/*!
+ * Simple performance tracing counter.
+ * \tparam T Arithmetic counter type
+ *
+ * Records a named value at the current timestamp which
+ * can then be displayed on a timeline.
+ *
+ * See https://perfetto.dev/docs/instrumentation/track-events#counters
+ */
 template<class T>
 inline void trace_counter(char const* name, T value)
 {
-    static_assert(std::is_arithmetic_v<T>, "Only support numeric counters");
-    if (use_profiling())
+    static_assert(std::is_arithmetic_v<T>,
+                  "Only numeric counters are supported");
+    if ((CELERITAS_USE_PERFETTO || CELERITAS_USE_CUDA) && use_profiling())
     {
-        // on some platform size_t is equivalent to uint64_t, which would cause
+        // On some platform size_t is equivalent to uint64_t, which would cause
         // duplicate template instantiation
         using counter_type = std::conditional_t<
             std::is_same_v<T, std::size_t>,
@@ -39,15 +46,6 @@ inline void trace_counter(char const* name, T value)
         detail::trace_counter_impl<counter_type>(name, value);
     }
 }
-#else
-
-// Ignore if Perfetto is unavailable
-template<class T>
-inline void trace_counter(char const*, T)
-{
-}
-
-#endif
 
 //---------------------------------------------------------------------------//
 }  // namespace celeritas
